PaymentStatus as interface #1053
