+++ commercetools-models/src/main/java/io/sphere/sdk/payments/PaymentStatus.java
@@ -1,7 +1,6 @@
 package io.sphere.sdk.payments;
 
-import com.fasterxml.jackson.annotation.JsonCreator;
-import io.sphere.sdk.models.Base;
+import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import io.sphere.sdk.models.Reference;
 import io.sphere.sdk.states.State;
 
@@ -10,33 +9,14 @@ import javax.annotation.Nullable;
 /**
  @see PaymentStatusBuilder
  */
-public final class PaymentStatus extends Base {
+@JsonDeserialize(as = PaymentStatusImpl.class)
+public interface PaymentStatus {
     @Nullable
-    private final String interfaceCode;
-    @Nullable
-    private final String interfaceText;
-    @Nullable
-    private final Reference<State> state;
-
-    @JsonCreator
-    PaymentStatus(@Nullable final String interfaceCode, @Nullable final String interfaceText, @Nullable final Reference<State> state) {
-        this.interfaceCode = interfaceCode;
-        this.interfaceText = interfaceText;
-        this.state = state;
-    }
+    String getInterfaceCode();
 
     @Nullable
-    public String getInterfaceCode() {
-        return interfaceCode;
-    }
+    String getInterfaceText();
 
     @Nullable
-    public String getInterfaceText() {
-        return interfaceText;
-    }
-
-    @Nullable
-    public Reference<State> getState() {
-        return state;
-    }
+    Reference<State> getState();
 }
